// front/src/providers/api-client.generated/model/anonymousExchangeDto.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AnonymousMessageSenderType } from './anonymousMessageSenderType';
import { UserDto } from './userDto';
import { AppFileDto } from './appFileDto';
import { CandidateApplicationDto } from './candidateApplicationDto';


export interface AnonymousExchangeDto { 
    id?: string;
    creationDate?: Date;
    modifDate?: Date;
    disabled?: boolean;
    candidateApplicationId: string;
    candidateApplication?: CandidateApplicationDto;
    consultantId?: string;
    consultant?: UserDto;
    messageContent: string;
    seen?: boolean;
    senderType?: AnonymousMessageSenderType;
    fileId?: string;
    file?: AppFileDto;
}