// front/src/providers/api-client.generated/model/anonymousMessageSenderType.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type AnonymousMessageSenderType = 'consultant' | 'guest';

export const AnonymousMessageSenderType = {
    Consultant: 'consultant' as AnonymousMessageSenderType,
    Guest: 'guest' as AnonymousMessageSenderType
};