// front/src/providers/api-client.generated/model/appFileDto.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AppFileDto { 
    id?: string;
    creationDate?: Date;
    modifDate?: Date;
    disabled?: boolean;
    name: string;
    size?: number;
    mimeType: string;
    physicalName?: string;
    externalId?: string;
    fileType?: string;
    externalFilePath?: string;
}