// front/src/providers/api-client.generated/model/candidateApplicationDto.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CandidateDepartmentDto } from './candidateDepartmentDto';
import { AppFileDto } from './appFileDto';
import { AddressDto } from './addressDto';
import { AnonymousExchangeDto } from './anonymousExchangeDto';
import { CandidateCountryDto } from './candidateCountryDto';
import { CandidateDto } from './candidateDto';
import { AppValueDto } from './appValueDto';
import { CandidateApplicationJobsDto } from './candidateApplicationJobsDto';


export interface CandidateApplicationDto { 
    id?: string;
    lastName: string;
    firstName: string;
    genderId: string;
    gender?: AppValueDto;
    birthDate: Date;
    partnerFirstName?: string;
    partnerLastName?: string;
    partnerGenderId?: string;
    partnerGender?: AppValueDto;
    partnerBirthDate?: Date;
    partnerEmail?: string;
    partnerPhone?: string;
    professionId?: string;
    profession?: AppValueDto;
    address?: AddressDto;
    phone: string;
    phoneSecondary: string;
    email: string;
    skills: string;
    relationshipStatusId: string;
    relationshipStatus?: AppValueDto;
    creationDate?: Date;
    modifDate?: Date;
    applyStatus?: AppValueDto;
    applyStatusId?: string;
    inCouple: boolean;
    spontaneousApplication: boolean;
    candidateId?: string;
    mainPhotoBase64?: string;
    mainPhotoBase64MimeType?: string;
    photoFileId?: string;
    photoFile?: AppFileDto;
    mainResumeFileId?: string;
    mainResumeFile?: AppFileDto;
    resumeFileBase64?: string;
    resumeFileBase64MimeType?: string;
    partnerResumeFileId?: string;
    partnerResumeFile?: AppFileDto;
    partnerResumeFileBase64?: string;
    partnerResumeFileBase64MimeType?: string;
    seen?: boolean;
    candidateApplicationJobs?: Array<CandidateApplicationJobsDto>;
    disabled?: boolean;
    jobOfferLinkedRef?: string;
    linkedToCandidate?: boolean;
    candidate?: CandidateDto;
    guidExchange?: string;
    anonymousExchanges?: Array<AnonymousExchangeDto>;
    newsletterUnsubscribed?: boolean;
    newsletterUnsubscribedGuid?: string;
    candidateDepartments?: Array<CandidateDepartmentDto>;
    candidateCountries?: Array<CandidateCountryDto>;
    allowed_to_work_us?: boolean;
    require_sponsorship_us?: boolean;
    usTermsAcceptedAt?: Date;
    usTermsVersion?: string;
}